refactor: rename models to modeling for impact objects
