+++ ecologits/tracers/anthropic_tracer.py
@@ -6,7 +6,7 @@ from typing import Any, Callable, Generic, Optional, TypeVar
 from typing_extensions import override
 from wrapt import wrap_function_wrapper
 
-from ecologits.impacts.models import Impacts
+from ecologits.impacts import Impacts
 from ecologits.tracers.utils import llm_impacts
 
 try:
