+++ ecologits/tracers/huggingface_tracer.py
@@ -5,7 +5,7 @@ from typing import Any, Callable, Union
 
 from wrapt import wrap_function_wrapper
 
-from ecologits.impacts.models import Impacts
+from ecologits.impacts import Impacts
 from ecologits.tracers.utils import llm_impacts
 
 try:
